CLASS SEPERATION
Class systems are hierarchies within a community that separate groups based on one of several factors. Traditionally in India, the class
system is dependent upon who one's parents are; and once born into this class, it is generally impossible to move up in the system. India's
traditional system is based on its own culture, which is grounded in Hinduism. However, class systems are not always tied to a culture.
Typically, class systems exist across entire countries, regardless of the cultures that different groups may have. They are also based on the
economic status of citizens. While being born into poverty puts people in a lower socioeconomic status than others, it is possible to gain
wealth and move up in this economic hierarchy.
